(PC-17881)[API] refactor: every `young_status`'s stuff to his own file
