+++ api/src/pcapi/core/users/young_status.py
@@ -1,39 +1,47 @@
-
 from datetime import datetime
+import enum
 
 import attrs
-from pcapi.core.users import constants
+
 from pcapi.core.users import models
+
+
+class YoungStatusType(enum.Enum):
+    ELIGIBLE = "eligible"
+    NON_ELIGIBLE = "non_eligible"
+    BENEFICIARY = "beneficiary"
+    EX_BENEFICIARY = "ex_beneficiary"
+    SUSPENDED = "suspended"
 
 
 @attrs.frozen
 class YoungStatus:
-    status_type: constants.YoungStatusType
+    status_type: YoungStatusType
 
 
 @attrs.frozen
 class Eligible(YoungStatus):
-    status_type: constants.YoungStatusType = constants.YoungStatusType.ELIGIBLE
+    status_type: YoungStatusType = YoungStatusType.ELIGIBLE
 
 
 @attrs.frozen
 class NonEligible(YoungStatus):
-    status_type: constants.YoungStatusType = constants.YoungStatusType.NON_ELIGIBLE
+    status_type: YoungStatusType = YoungStatusType.NON_ELIGIBLE
 
 
 @attrs.frozen
 class Beneficiary(YoungStatus):
-    status_type: constants.YoungStatusType = constants.YoungStatusType.BENEFICIARY
+    status_type: YoungStatusType = YoungStatusType.BENEFICIARY
 
 
 @attrs.frozen
 class ExBeneficiary(YoungStatus):
-    status_type: constants.YoungStatusType = constants.YoungStatusType.EX_BENEFICIARY
+    status_type: YoungStatusType = YoungStatusType.EX_BENEFICIARY
 
 
 @attrs.frozen
 class Suspended(YoungStatus):
-    status_type: constants.YoungStatusType = constants.YoungStatusType.SUSPENDED
+    status_type: YoungStatusType = YoungStatusType.SUSPENDED
 
 
 def young_status(user: models.User) -> YoungStatus:
